[1] Bumped major number because of (probably benign) API change. [1.1] Changed a long to an int; really a bug. [2] Minor refactoring for API change to conform to fcntl(2) prototype. [3] unittest-ipc PASSED. [4] unittest-ipc4 PASSED.

+++ Diminuto/inc/com/diag/diminuto/diminuto_ipc.h
@@ -4,7 +4,7 @@
 
 /**
  * @file
- * @copyright Copyright 2010-2024 Digital Aggregates Corporation, Colorado, USA.
+ * @copyright Copyright 2010-2025 Digital Aggregates Corporation, Colorado, USA.
  * @note Licensed under the terms in LICENSE.txt.
  * @brief Provides capabilities common to IPv4, IPv6, and local sockets.
  * @author Chip Overclock <mailto:coverclock@diag.com>
@@ -177,7 +177,7 @@ extern int diminuto_ipc_set_interface(int fd, const char * interface);
  * @param mask is a bit mask indication the option to enable or disable.
  * @return >=0 for success, <0 for error.
  */
-extern int diminuto_ipc_set_status(int fd, int enable, long mask);
+extern int diminuto_ipc_set_status(int fd, int enable, int mask);
 
 /**
  * Set a socket option (setsockopt(2)) on a socket.
